Calculate phi assigns
This basically means the following. Given:

    %<dest> = phi <bb0> => %<value0>, <bb1> => %<value1>, ...

Add the following assignments:

    bb0:
      %<dest> <= %<value0>

    bb1:
      %<dest> <= %<value1>
...

This will simplify the interpreter. Instead of having to track the last
executed basic block, we just add these assignments before the final
branch at the end of the basic block.

This also will simplify the bytecode generator. When generating code for
a basic block, we know which instructions must be added without having
to look it up in the successor basic blocks.