Handle 64-bit values in the bitstream properly
It seems that 64-bit values only really occur as constants and switch
case values. This means we can use 32-bit values for most places.

One weird thing is that the signed encoding LLVM uses requires INT32_MIN
to be encoded as a 64-bit value (0x100000001). So we read 64-bit values
any time pn_decode_sign_rotated_value is used.

Also all VBRs are decoded as int64 internally now. It is necessary
because any sign-extended values will be sign-extended to int64. In most
cases we check that the value can be truncated to 32-bit without losing
data and use that instead.