Implement SETJMP/LONGJMP intrinsics
The implementation is cheesy, but seems to work.

Basically, we assign a rolling ID whenever setjmp is called. This ID and
the current frame info are stored in the PNJmpBuf structure, which is
stored as a link list in PNCallFrame. This ID is also written into the
jmp_buf struct.

If the function returns without calling longjmp, the allocated data is
thrown away (just like a normal call frame). If longjmp is called, we
read the ID from the jmp_buf, and iterate through the call stack looking
for the matching jmpbuf id. If we find it, then kill all the
  intermediate frames and restore the stored call frame.