Simple fix for phi assignment ordering problem
Phi nodes are assumed to be assigned all-at-once; that is, the source
values are read from the previous basic block and written to the
destination values. Since we perform the assignments piecemeal, the
ordering matters. For example,

  a' = b
  c' = a

This assigns the same value to a' and c' (b). One way to fix this is to
reorder the assignments, though that won't always work.

This CL fixes the issue by always using temporary storage for phi
assignments. A better solution may be to topologically sort the
dependencies.