Execute from an instruction stream
This is a relatively mechanical change to flatten the instructions from
each basic block into one stream per function. The instructions are
currently duplicated in the basic block instructions list as well, but
that will be removed in a later CL.

Because the PNInstruction is used for reading and for execution, the
bb_id members are reused as an instruction id. That will be cleaned up
later too.