Allocate function stuff from temp_allocator
The function stuff (instructions, basic blocks, etc.) is not used after
the instruction stream has been written. If we allocate it with the
temp_allocator, we can throw it away after the function has been read.

There's still a few things that need to be removed as well; function
values and constants can be removed too, but not until that info is also
written to the instruction stream.