Simplify exit behavior
Always set the thread state to PN_THREAD_DEAD when exiting from a
thread, whether main thread or not. This way we only have to check the
thread state in the instruction quantum loop.

Also, only set executor->exiting when returning from the main thread,
not from any other thread.