More work on ppapi, implemented an event loop
This works by making executor->main_thread a pointer, and swapping out
the main "thread" for an event loop "thread".

The primary difference is that the event loop thread has a call stack
limited to where PPAPI_START was called. When the call stack is emptied,
the event is finished, and the next event in the queue is processed.