Implement fast past for phi assigns
Added a quick check that no source depends on a previous dest in phi
assignments. If that's true, we can write do the assignments directly.
If not, it falls back on copying everything to temporaries.

Also, I moved the phi assign tracing out of the executor and into
pn_trace.h